moving workloads/imagenet to workloads/imagenet_resnet/jax and reference_submissions/imagenet to reference_submissions/imagenet_resnet
