+++ algorithmic_efficiency/workloads/cifar/cifar_jax/workload.py
@@ -13,7 +13,7 @@ from algorithmic_efficiency import param_utils
 from algorithmic_efficiency import spec
 from algorithmic_efficiency.workloads.cifar.cifar_jax import input_pipeline
 from algorithmic_efficiency.workloads.cifar.workload import BaseCifarWorkload
-from algorithmic_efficiency.workloads.imagenet.imagenet_jax import models
+from algorithmic_efficiency.workloads.imagenet_resnet.imagenet_jax import models
 
 
 class CifarWorkload(BaseCifarWorkload):
