+++ algorithmic_efficiency/workloads/imagenet_resnet/imagenet_jax/models.py
