+++ algorithmic_efficiency/workloads/imagenet_resnet/imagenet_jax/workload.py
@@ -12,10 +12,10 @@ import tensorflow_datasets as tfds
 
 from algorithmic_efficiency import param_utils
 from algorithmic_efficiency import spec
-from algorithmic_efficiency.workloads.imagenet.imagenet_jax import \
+from algorithmic_efficiency.workloads.imagenet_resnet.imagenet_jax import \
     input_pipeline
-from algorithmic_efficiency.workloads.imagenet.imagenet_jax import models
-from algorithmic_efficiency.workloads.imagenet.workload import \
+from algorithmic_efficiency.workloads.imagenet_resnet.imagenet_jax import models
+from algorithmic_efficiency.workloads.imagenet_resnet.workload import \
     BaseImagenetWorkload
