+++ reference_submissions/imagenet_resnet/imagenet_jax/submission.py
