+++ tests/test_param_shapes.py
@@ -2,9 +2,9 @@ import jax
 import numpy as np
 import pytest
 
-from algorithmic_efficiency.workloads.imagenet.imagenet_jax.workload import \
+from algorithmic_efficiency.workloads.imagenet_resnet.imagenet_jax.workload import \
     ImagenetWorkload as JaxImagenetWorkload
-from algorithmic_efficiency.workloads.imagenet.imagenet_pytorch.workload import \
+from algorithmic_efficiency.workloads.imagenet_resnet.imagenet_pytorch.workload import \
     ImagenetWorkload as PyTorchImagenetWorkload
 from algorithmic_efficiency.workloads.mnist.mnist_jax.workload import \
     MnistWorkload as JaxMnistWorkload
