+++ tests/workloads/imagenet/imagenet_jax/workload_test.py
@@ -1,11 +1,11 @@
-"""Tests for imagenet_jax/workload.py."""
+"""Tests for resnet/jax/workload.py."""
 
 from absl.testing import absltest
 import jax
 import jax.numpy as jnp
 
 from algorithmic_efficiency import spec
-from algorithmic_efficiency.workloads.imagenet.imagenet_jax.workload import \
+from algorithmic_efficiency.workloads.imagenet_resnet.imagenet_jax.workload import \
     ImagenetWorkload
 
 
@@ -16,7 +16,7 @@ def _pytree_total_diff(pytree_a, pytree_b):
 
 
 class ModelsTest(absltest.TestCase):
-  """Tests for imagenet_jax/workload.py."""
+  """Tests for resnet/jax/workload.py."""
 
   def test_forward_pass(self):
     batch_size = 11
